Safari-609.1.11.1/LayoutTests/media/modern-media-controls/tracks-panel/tracks-panel-prevent-default-on-keydown-expected.txt

Testing a TracksPanel calls preventDefault() on events it handles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


mediaControls.showTracksPanel()

Pressing the ArrowDown key
PASS event.defaultPrevented is true
Pressing the ArrowRight key
PASS event.defaultPrevented is false

PASS successfullyParsed is true

TEST COMPLETE
